Rewrite parser.
Rewrite the parser due to not begin able to change reduction that is a
order in Happy. Upon encountering a lambda token (`\`) a monadic action
adding the bound variables to the local context should be executed.
This is not possible in Happy since it uses depth-first reduction, i.e.
the body is already reduced by the time bound variables are.