Rewrite parser
Rewrite the parser due to not begin able to change reduction order in
Happy. Upon encountering a lambda token (`\`) a monadic action adding
the bound variables to the local context should be executed. This is not
possible with Happy since it uses depth-first reduction, i.e. the body
is is already reduced by the time the lambda binder is.